Extract the MultiList component
